+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Removed
+- Embeds from the codebase, left with a message to migrate to the new CDN
 
 ## [4.44.0][] - 2017-10-19
 ### Added